reco/requirements.txt
psycopg2-binary>=2.9,<3
python-dateutil>=2.8,<3


